Create a package for a Plumber API using plungr

Create a package for a Plumber API using plungr

create_plungr(
  path,
  check_name = TRUE,
  open = TRUE,
  package_name = basename(path)
)

Arguments

path
Name of the folder to create the package in. This will also be used as the package name.

check_name
When using this function in the console, you can prevent the package name from being checked.

open
Boolean open the created project

package_name
Package name to use by default it's `basename(path)` but if path == '.' and `package_name` not explicitly given, then `basename(getwd())` will be used.